Initial Notebook Upload
In this initial upload, I introduced the project, uploaded the files, and created sections where I started cleaning and preparing the data for the rest of the project. I merged the 4 data frames into a single data frame and identified a list of tasks I needed to do to continue to clean and prepare the data.